Fix: fix the error about navigation and create bottombar composable
